% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_sentiments_function.R
\name{count_sentiments}
\alias{count_sentiments}
\title{Allows to graphically represent the positive and negative words a speaker says.}
\usage{
count_sentiments(data, ...)
}
\arguments{
\item{data}{dataset}

\item{...}{vector containing strings}
}
\description{
Allows to graphically represent the positive and negative words a speaker says.
}
\details{
Using the "bing" sentiment codes, allows you to count and graph the number of positive and negative words a 
speaker will say.
}
\examples{
count_sentiments(eleventh, "Eleventh:")
}
