% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plot_counts}
\alias{plot_counts}
\title{Create bar charts for speaker counts}
\usage{
plot_counts(data, ...)
}
\description{
Create bar charts for speaker counts
}
\details{
For multiple speakers, it is important to contain the names beteen "c()".
See examples for details.
}
\examples{
plot_counts(data, "ROSE:")
plot_counts(data, c("Eleventh:", "CLARA:", "RIVER:"))
}
