% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_script_function.R
\name{tidy_script}
\alias{tidy_script}
\title{Edit data that has been scrapped from the web}
\usage{
tidy_script(doctor_text, Doctor)
}
\description{
Edit data that has been scrapped from the web
}
\details{
It is important to have the doctor's name spelled the way you would like it presented within 
the data. Allows you to clean datasets pulled from the web so that each line is it's own row and
that the speaker and dialogue have their own column
}
\examples{
tidy_script(eleventh, "Eleventh:")
}
