% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscraping_function.R
\name{webscrape_lines}
\alias{webscrape_lines}
\title{Pull scripts from online}
\usage{
webscrape_lines(url)
}
\description{
Pull scripts from online
}
\details{
This function allows you to get the doctor who lines form the website: wwww.chakoteya.net
}
\examples{
url <- "http://www.chakoteya.net/DoctorWho/episodes11.html"
webscrape_lines(url)
}
